+++ yt_dlp/utils/_deprecated.py
@@ -1,6 +1,6 @@
 """Deprecated - New code should avoid these"""
 
-from ._old import preferredencoding
+from ._utils import preferredencoding
 
 
 def encodeFilename(s, for_subprocess=False):
