+++ yt_dlp/utils/_utils.py
